chore: move gpu types outside of build tags
